Move and refactor into hooks, and connect - test successful. MVP Completed
