+++ src/components/Navbar.jsx
@@ -1,34 +1,15 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
+import assignClassName from "./Hooks/darkMode";
+import useLocalStorage from "./Hooks/useLocalStorage";
 
 const Navbar = () => {
-  const [darkMode, setDarkMode] = useState(false);
+  const [darkMode, setDarkMode] = useLocalStorage("Dark-Mode", false);
+  assignClassName(darkMode);
 
   const toggleMode = e => {
     e.preventDefault();
     setDarkMode(!darkMode);
   };
-
-  // darkMode = value to be set to the local storage
-  const useLocalStorage = (key, initialValue) => {
-    const [storedValue, setStoredValue] = useState(() => {
-      const keyValue = window.localStorage.getItem(key);
-      return keyValue ? JSON.parse(keyValue) : initialValue;
-    });
-    const setKeyValue = value => {
-      setStoredValue(value);
-      window.localStorage.setItem(key, JSON.stringify(value));
-    };
-    return [storedValue, setKeyValue];
-  };
-
-  useEffect(() => {
-    // if darkMode = true, then assign class name dark-mode
-    if (darkMode === true) {
-      document.querySelector(`body`).classList.add("dark-mode");
-    } else {
-      document.querySelector(`body`).classList.remove("dark-mode");
-    }
-  });
 
   return (
     <nav className="navbar">
